backend/requirements.txt
fastapi>=0.104.0
uvicorn==0.34.3
gunicorn==21.2.0
python-multipart==0.0.9
python-dotenv==1.0.0
groq>=0.8.0
pinecone-client==3.2.1
sentence-transformers==2.7.0
pypdf==4.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
pydantic>=2.0.0,<3.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0
python-dateutil>=2.8.2
